emoi backend project code/emoi_backend_project/requirements.txt
pandas
boto3==1.34.14
botocore==1.34.14
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
cssselect2==0.7.0
Deprecated==1.2.14
distro==1.9.0
Django==4.1.13
django-cors-headers==4.3.1
django-cron==0.6.0
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1

# mysql-connector-python==8.4.0
numpy==1.26.3
opencv-python==4.10.0.84
openpyxl==3.1.2
reportlab==4.2.2
requests==2.31.0

gunicorn
